DRRA-Agent/requirements.txt
# Core dependencies for DRRA-Agent
python-dateutil
dataclasses-json



google-generativeai